# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LibraryOS API - Multi-tenant school bookstore management platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login (tenant admin, cashier, super admin, or student)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register-student:
    post:
      operationId: registerStudent
      tags: [auth]
      summary: Register a student account for online store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterStudentRequest"
      responses:
        "201":
          description: Student registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ========== TENANTS (Super Admin) ==========
...
  /cards:
    get:
      operationId: listCards
      tags: [cards]
      summary: List physical platform cards for current tenant
      responses:
        "200":
          description: List of cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
    post:
      operationId: createCard
      tags: [cards]
      summary: Add physical card stock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardRequest"
      responses:
        "201":
          description: Card created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /cards/{cardId}:
    patch:
      operationId: updateCard
      tags: [cards]
      summary: Update card
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCardRequest"
      responses:
        "200":
          description: Card updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
    delete:
      operationId: deleteCard
      tags: [cards]
      summary: Delete card
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ========== EDUCATIONAL PLATFORMS (Super Admin) ==========
...
  /discounts:
    get:
      operationId: listDiscounts
      tags: [discounts]
      summary: List discounts for current tenant
      responses:
        "200":
          description: List of discounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Discount"
    post:
      operationId: createDiscount
      tags: [discounts]
      summary: Create a discount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDiscountRequest"
      responses:
        "201":
          description: Discount created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discount"
  /discounts/{discountId}:
    patch:
      operationId: updateDiscount
      tags: [discounts]
      summary: Update a discount
      parameters:
        - name: discountId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDiscountRequest"
      responses:
        "200":
          description: Discount updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discount"
    delete:
      operationId: deleteDiscount
      tags: [discounts]
      summary: Delete a discount
      parameters:
        - name: discountId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /discounts/{discountId}/toggle:
    post:
      operationId: toggleDiscount
      tags: [discounts]
      summary: Toggle discount active/inactive
      parameters:
        - name: discountId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discount"

  # ========== PRINT SERVICES ==========
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [super_admin, tenant_admin, cashier, student]
      required:
        - username
        - password
        - role
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserInfo"
      required:
        - token
        - user
    UserInfo:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [super_admin, tenant_admin, cashier, student]
        tenantId:
          type: ["integer", "null"]
        tenantName:
          type: ["string", "null"]
        tenantSlug:
          type: ["string", "null"]
      required:
        - id
        - name
        - username
        - role
    RegisterStudentRequest:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        password:
          type: string
        tenantSlug:
          type: string
      required:
        - fullName
        - phone
        - password
        - tenantSlug

    # TENANT
...
    Card:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        platformName:
          type: string
        cardValue:
          type: number
        quantity:
          type: integer
        price:
          type: number
        alertThreshold:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - platformName
        - cardValue
        - quantity
        - price
        - alertThreshold
        - createdAt
    CreateCardRequest:
      type: object
      properties:
        platformName:
          type: string
        cardValue:
          type: number
        quantity:
          type: integer
        price:
          type: number
        alertThreshold:
          type: integer
      required:
        - platformName
        - cardValue
        - quantity
        - price
        - alertThreshold
    UpdateCardRequest:
      type: object
      properties:
        quantity:
          type: integer
        price:
          type: number
        alertThreshold:
          type: integer

    # PLATFORM (Educational)
...
    Discount:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [product, category, buy_x_get_y, invoice]
        discountValue:
          type: number
        discountType:
          type: string
          enum: [percent, amount]
        productId:
          type: ["integer", "null"]
        productName:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        buyQuantity:
          type: ["integer", "null"]
        getFreeQuantity:
          type: ["integer", "null"]
        minInvoiceAmount:
          type: ["number", "null"]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - name
        - type
        - discountValue
        - discountType
        - startDate
        - endDate
        - isActive
        - createdAt
    CreateDiscountRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [product, category, buy_x_get_y, invoice]
        discountValue:
          type: number
        discountType:
          type: string
          enum: [percent, amount]
        productId:
          type: integer
        category:
          type: string
        buyQuantity:
          type: integer
        getFreeQuantity:
          type: integer
        minInvoiceAmount:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isActive:
          type: boolean
      required:
        - name
        - type
        - discountValue
        - discountType
        - startDate
        - endDate
    UpdateDiscountRequest:
      type: object
      properties:
        name:
          type: string
        discountValue:
          type: number
        discountType:
          type: string
          enum: [percent, amount]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isActive:
          type: boolean

    # PRINT SERVICE
...