# lib/api-zod/src/generated/types/loginRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LibraryOS API - Multi-tenant school bookstore management platform
 * OpenAPI spec version: 0.1.0
 */
import type { LoginRequestRole } from "./loginRequestRole";

export interface LoginRequest {
  username: string;
  password: string;
  role: LoginRequestRole;
}